chore: fix all archteture
