+++ lib/payments_handler/db/engines/events_list.ex
@@ -1,6 +1,6 @@
-defmodule PaymentsHandler.Db.EventsList do
-  alias PaymentsHandler.Db.EventsServer.ListStrategy
-  alias PaymentsHandler.Payments.Events
+defmodule PaymentsHandler.Db.Engines.EventsList do
+  alias PaymentsHandler.Db.Engines.Strategies.ListStrategy
+  alias PaymentsHandler.Db.Engines.Objects.Events
 
   defdelegate construct, to: ListStrategy, as: :construct
   defdelegate get_from_id(tree, index), to: ListStrategy, as: :get_from_id
