+++ lib/payments_handler/db/engines/strategy_behavior/db_engine_behaviour.ex
@@ -1,0 +1,5 @@
+defmodule PaymentsHandler.Db.StrategyBehavior.DbEngine do
+  @callback construct() :: struct()
+  @callback get_from_id(tree :: struct(), index :: term) :: {index :: term, values :: term} | nil
+  @callback insert(tree :: struct(), index :: term, item :: term) :: struct()
+end
